adjusted endpoints
I was pushing req.params.movieId directly into the FavoriteMovies array. In my ProfileView component on the frontend, I was sending an array of favorite movie IDs in the request body when adding and removing favorite movies, updated this endpoint to accept JSON data in the request body instead of using route parameters.