adjusted use of fetch
I moved the fetch request inside a useEffect hook. This way, the fetch request is executed only when certain dependencies (in this case, user and token) change. and I used async/await to make the fetch request asynchronous. This ensures that the component doesn't proceed to the next steps until the fetch operation has completed